refactor(oauth): split token route into lib/oauth/grants/ modules
- Extract handleDeviceCodeGrant → lib/oauth/grants/device-code.ts
- Extract handleRefreshTokenGrant → lib/oauth/grants/refresh.ts
- Extract mintTokensResponse + secondsUntil + new graceRetryResponse helper
  → lib/oauth/grants/mint.ts
- Dedupe the two grace-window blocks via graceRetryResponse
- Flatten device-code status guards into an exhaustive switch with handlers
  per status (handlePendingPoll / handleApproved)
- Replace IIFE in mintTokensResponse with a plain conditional
- Token route is now a ~25-line dispatcher matching the rest of lib/oauth/'s
  factoring style
- Add TODO(rate-limit) on /oauth/device/authorize covering the deferred
  rate-limiting infrastructure for the OAuth surface
